pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lmr"
version = "0.1.0"
description = "Research code for training large language models."
readme = "README.md"
requires-python = ">=3.9"

# Python dependencies
dependencies = [
    # Core
    "torch==2.7.0",
    "torchtune==0.6.1",
    "torchao==0.11.0",
    "datasets==3.6.0",
    "transformers==4.52.4",
    "huggingface-hub==0.31.4",

    # Config
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",

    # Util
    "tqdm==4.67.1",
    "zstandard>=0.21.0,<1.0.0",
    "python-dotenv==1.1.1"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["lmr"]
